+++ core/src/main/scala/commbank/coppersmith/Feature.scala
@@ -54,6 +54,7 @@ object Feature {
     case class Str(value: Option[String])           extends Value
     case class Date(value: Option[Datestamp])       extends Value
     case class Time(value: Option[Timestamp])       extends Value
+    case class Bool(value: Option[Boolean])         extends Value
 
     implicit def fromInt(i: Int):                         Integral      = Option(i)
     implicit def fromLong(l: Long):                       Integral      = Option(l)
@@ -62,6 +63,7 @@ object Feature {
     implicit def fromString(s: String):                   Str           = Option(s)
     implicit def fromDate(d: Datestamp):                  Date          = Option(d)
     implicit def fromTime(t: Timestamp):                  Time          = Option(t)
+    implicit def fromBoolean(b: Boolean):                 Bool          = Option(b)
     implicit def fromOInt(i: Option[Int]):                Integral      = Integral(i.map(_.toLong))
     implicit def fromOLong(l: Option[Long]):              Integral      = Integral(l)
     implicit def fromODouble(d: Option[Double]):          FloatingPoint = FloatingPoint(d)
@@ -69,6 +71,7 @@ object Feature {
     implicit def fromOString(s: Option[String]):          Str           = Str(s)
     implicit def fromODate(d: Option[Datestamp]):         Date          = Date(d)
     implicit def fromOTime(t: Option[Timestamp]):         Time          = Time(t)
+    implicit def fromOBoolean(b: Option[Boolean]):        Bool          = Bool(b)
 
     implicit val intOrder: Order[Integral]      = orderBy(_.value)
     implicit val decOrder: Order[Decimal]       = orderBy(_.value)
@@ -77,7 +80,7 @@ object Feature {
     implicit val dOrder:   Order[Date]          = orderBy(_.value)
     implicit val tOrder:   Order[Time]          = orderBy(_.value)
 
-    abstract class Range[+V : Order] {
+    sealed abstract class Range[+V : Order] {
       // V needs to be covariant to satisfy Metadata type constraint, so can't be in contravariant
       // position here. This problem goes away when switching to arbitrary value types.
       // def contains(v: V): Boolean
@@ -95,17 +98,18 @@ object Feature {
     }
     object SetRange {
       // Should return Range[V] once V is made invariant on Range and contains is added back
-      def apply[V : Order](values: List[V]): SetRange[V] = SetRange(ListSet(values: _*))
+      def apply[V : Order](values: V*): SetRange[V] = SetRange(ListSet(values: _*))
     }
   }
 
   // Legal type/value combinations
   @implicitNotFound("Features with value type ${V} cannot be ${T}")
   abstract class Conforms[T <: Type : TypeTag, V <: Value : TypeTag] {
-    def typeTag:  TypeTag[T] = implicitly
-    def valueTag: TypeTag[V] = implicitly
+    def typeTag: TypeTag[T] = implicitly
+    def valueType: Metadata.ValueType = Metadata.valueType[V]
   }
   implicit object NominalStr              extends Conforms[Type.Nominal.type,    Value.Str]
+  implicit object OrdinalStr              extends Conforms[Type.Ordinal.type,    Value.Str]
 
   implicit object OrdinalDecimal          extends Conforms[Type.Ordinal.type,    Value.Decimal]
   implicit object ContinuousDecimal       extends Conforms[Type.Continuous.type, Value.Decimal]
@@ -113,20 +117,22 @@ object Feature {
   implicit object OrdinalFloatingPoint    extends Conforms[Type.Ordinal.type,    Value.FloatingPoint]
   implicit object ContinuousFloatingPoint extends Conforms[Type.Continuous.type, Value.FloatingPoint]
 
+  implicit object NominalIntegral         extends Conforms[Type.Nominal.type,    Value.Integral]
   implicit object OrdinalIntegral         extends Conforms[Type.Ordinal.type,    Value.Integral]
   implicit object ContinuousIntegral      extends Conforms[Type.Continuous.type, Value.Integral]
-
   implicit object DiscreteIntegral        extends Conforms[Type.Discrete.type,   Value.Integral]
 
   implicit object InstantDate             extends Conforms[Type.Instant.type,    Value.Date]
   implicit object InstantTime             extends Conforms[Type.Instant.type,    Value.Time]
 
+  implicit object NominalBool             extends Conforms[Type.Nominal.type,    Value.Bool]
+
   object Conforms {
 
-    def conforms_?[V <: Value : TypeTag](conforms: Conforms[_, _], metadata: Metadata[_, _]) = {
+    def conforms_?(conforms: Conforms[_, _], metadata: Metadata[_, _]) = {
       def getClazz(tag: TypeTag[_]) = tag.mirror.runtimeClass(tag.tpe.typeSymbol.asClass)
       metadata.featureType.getClass == getClazz(conforms.typeTag) &&
-        metadata.valueType == Metadata.valueType[V]
+        metadata.valueType == conforms.valueType
     }
   }
 
@@ -152,6 +158,7 @@ object Feature {
         case t if t =:= typeOf[Value.Decimal]       => ValueType.DecimalType
         case t if t =:= typeOf[Value.FloatingPoint] => ValueType.FloatingPointType
         case t if t =:= typeOf[Value.Str]           => ValueType.StringType
+        case t if t =:= typeOf[Value.Bool]          => ValueType.BoolType
         case t if t =:= typeOf[Value.Date]          => ValueType.DateType
         case t if t =:= typeOf[Value.Time]          => ValueType.TimeType
       }
@@ -162,6 +169,7 @@ object Feature {
       case object DecimalType       extends ValueType
       case object FloatingPointType extends ValueType
       case object StringType        extends ValueType
+      case object BoolType          extends ValueType
       case object DateType          extends ValueType
       case object TimeType          extends ValueType
     }
@@ -208,13 +216,13 @@ abstract class Feature[S, +V <: Value](val metadata: Metadata[S, V]) {
 }
 
 case class FeatureValue[+V <: Value](
-  entity:  EntityId,
-  name:    Name,
-  value:   V
+  entity: EntityId,
+  name:   Name,
+  value:  V
 )
 
 object FeatureValue {
-  implicit class AsEavt[V <: Value](fv: FeatureValue[V]) {
-    def asEavt(time: FeatureTime): (EntityId, Name, V, FeatureTime) = (fv.entity, fv.name, fv.value, time)
+  implicit class AsEavt(fv: FeatureValue[Value]) {
+    def asEavt(time: FeatureTime): (EntityId, Name, Value, FeatureTime) = (fv.entity, fv.name, fv.value, time)
   }
 }