+++ core/src/test/scala/commbank/coppersmith/Arbitraries.scala
@@ -14,8 +14,10 @@
 
 package commbank.coppersmith
 
+import scala.reflect._
 import scala.util.Try
 
+import scalaz.scalacheck.ScalazArbitrary._
 import scalaz.{Value => _, _}, Scalaz.{option => _, _}
 import scalaz.scalacheck.ScalaCheckBinding._
 
@@ -70,16 +72,110 @@ object Arbitraries {
     date <- arbLocalDate
   } yield Datestamp(date.getYear, date.getMonthOfYear, date.getDayOfMonth)
 
-  implicit val integralValueGen: Gen[Integral] = arbitrary[Option[Long]].map(Integral(_))
+  implicit val integralValueGen: Gen[Integral] = arbitrary[Option[Long]].map(Integral)
   implicit val decimalValueGen: Gen[Decimal] =
     arbitrary[Option[BigDecimal]].retryUntil(obd =>
-      Try(obd.hashCode()).isSuccess
-    ).map(Decimal(_))
-  implicit val floatingPointValueGen: Gen[FloatingPoint] = arbitrary[Option[Double]].map(FloatingPoint(_))
-  implicit val strValueGen: Gen[Str] = arbitrary[Option[String]].map(Str(_))
-  implicit val dateValueGen: Gen[Date] = arbitrary[Option[Datestamp]].map(Date(_))
-  implicit val timeValueGen: Gen[Time] = arbitrary[Option[Timestamp]].map(Time(_))
-  implicit val arbValue: Arbitrary[Value] = Arbitrary(oneOf(integralValueGen, decimalValueGen, floatingPointValueGen, strValueGen, dateValueGen, timeValueGen))
+      Try(obd.hashCode).isSuccess
+    ).map(Decimal)
+  implicit val floatingPointValueGen: Gen[FloatingPoint] = arbitrary[Option[Double]].map(FloatingPoint)
+  implicit val strValueGen: Gen[Str] = arbitrary[Option[String]].map(Str)
+  implicit val boolValueGen: Gen[Bool] = arbitrary[Option[Boolean]].map(Bool)
+  implicit val dateValueGen: Gen[Date] = arbitrary[Option[Datestamp]].map(Date)
+  implicit val timeValueGen: Gen[Time] = arbitrary[Option[Timestamp]].map(Time)
+
+  implicit val arbValue: Arbitrary[Value] = Arbitrary(oneOf(integralValueGen, decimalValueGen, floatingPointValueGen, strValueGen, boolValueGen, dateValueGen, timeValueGen))
+
+  // Generates values of the same subtype, but arbitrarily chooses the subtype to generate
+  implicit def arbValues: Arbitrary[NonEmptyList[Value]] =
+    Arbitrary(
+      oneOf(
+        NonEmptyListArbitrary(Arbitrary(decimalValueGen)).arbitrary,
+        NonEmptyListArbitrary(Arbitrary(floatingPointValueGen)).arbitrary,
+        NonEmptyListArbitrary(Arbitrary(integralValueGen)).arbitrary,
+        NonEmptyListArbitrary(Arbitrary(strValueGen)).arbitrary,
+        NonEmptyListArbitrary(Arbitrary(boolValueGen)).arbitrary,
+        NonEmptyListArbitrary(Arbitrary(dateValueGen)).arbitrary,
+        NonEmptyListArbitrary(Arbitrary(timeValueGen)).arbitrary
+      )
+    )
+
+  // Only for use with arbValues above - assumed values to be compared are of the same subtype
+  implicit val valueOrder: Order[Value] =
+    Order.order((a, b) => (a, b) match {
+      case (Decimal(d1), Decimal(d2)) => d1.cmp(d2)
+      case (FloatingPoint(d1), FloatingPoint(d2)) => d1.cmp(d2)
+      case (Integral(i1), Integral(i2)) => i1.cmp(i2)
+      case (Str(s1), Str(s2)) => s1.cmp(s2)
+      case (Bool(b1), Bool(b2)) => b1.cmp(b2)
+      case (Date(d1), Date(d2)) => d1.cmp(d2)
+      case (Time(t1), Time(t2)) => t1.cmp(t2)
+      case _ => sys.error("Assumption failed: Expected same value types from arbValues")
+    })
+
+  implicit val arbMinMaxRange: Arbitrary[MinMaxRange[Value]] = for {
+    vs <- arbValues
+  } yield MinMaxRange(vs.minimum1, vs.maximum1)
+
+  implicit val arbSetRange: Arbitrary[SetRange[Value]] = for {
+    vs <- arbValues
+  } yield SetRange(vs.toList:_*)
+
+  implicit val arbRange: Arbitrary[Option[Range[Value]]] =
+    Arbitrary(Gen.option(oneOf(arbMinMaxRange.arbitrary, arbSetRange.arbitrary)))
+
+  def typeMatches(c: Class[_], r: Option[Range[Value]]): Boolean = {
+    r match {
+      case Some(MinMaxRange(min, _)) if min.getClass == c => true
+      case Some(SetRange(vs)) if vs.isEmpty || vs.head.getClass == c => true
+      case None => true
+      case _ => false
+    }
+  }
+
+  def arbRangeOf[V <: Value : ClassTag]: Arbitrary[Option[Range[V]]] =
+    Arbitrary(arbRange.arbitrary.retryUntil(typeMatches(classTag[V].runtimeClass, _))
+      .map(_.asInstanceOf[Option[Range[V]]]))
+
+  implicit val arbStrRange: Arbitrary[Option[Range[Str]]] = arbRangeOf[Str]
+  implicit val arbIntegralRange: Arbitrary[Option[Range[Integral]]] = arbRangeOf[Integral]
+  implicit val arbFloatingPointRange: Arbitrary[Option[Range[FloatingPoint]]] = arbRangeOf[FloatingPoint]
+  implicit val arbDecimalRange: Arbitrary[Option[Range[Decimal]]] = arbRangeOf[Decimal]
+  implicit val arbDateRange: Arbitrary[Option[Range[Date]]] = arbRangeOf[Date]
+  implicit val arbTimeRange: Arbitrary[Option[Range[Time]]] = arbRangeOf[Time]
+
+  sealed trait RangeFieldPair {
+    def range: Option[Range[Value]]
+    def field: Field[Customer, _]
+  }
+  case class StrRangeFieldPair(range: Option[Range[Str]],
+                               field: Field[Customer, String]) extends RangeFieldPair
+  case class IntegralRangeFieldPair(range: Option[Range[Integral]],
+                                    field: Field[Customer, Int]) extends RangeFieldPair
+  case class FloatingPointRangeFieldPair(range: Option[Range[FloatingPoint]],
+                                         field: Field[Customer, Double]) extends RangeFieldPair
+
+  implicit val arbStrRangeFieldPair: Arbitrary[StrRangeFieldPair] = for {
+    r <- arbStrRange
+    f = Fields[Customer].Name
+  } yield StrRangeFieldPair(r, f)
+
+  implicit val arbIntegralRangeFieldPair: Arbitrary[IntegralRangeFieldPair] = for {
+    r <- arbIntegralRange
+    f = Fields[Customer].Age
+  } yield IntegralRangeFieldPair(r, f)
+
+  implicit val arbFloatingPointRangeFieldPair: Arbitrary[FloatingPointRangeFieldPair] = for {
+    r <- arbFloatingPointRange
+    f = Fields[Customer].Height
+  } yield FloatingPointRangeFieldPair(r, f)
+
+  implicit val arbRangeFieldPair: Arbitrary[RangeFieldPair] = {
+    Arbitrary(Gen.oneOf(
+      arbStrRangeFieldPair.arbitrary,
+      arbIntegralRangeFieldPair.arbitrary,
+      arbFloatingPointRangeFieldPair.arbitrary
+    ))
+  }
 
   val arbTimeMillis: Gen[Long] = arbitrary[DateTime].map(_.getMillis)
 
@@ -97,7 +193,8 @@ object Arbitraries {
        ageGen |@|
        arbitrary[Double] |@|
        arbitrary[Option[Double]] |@|
-       arbTimeMillis)(Customer.apply)
+       arbTimeMillis |@|
+       arbitrary[Boolean])(Customer.apply)
 
   implicit val arbCustomer: Arbitrary[Customer] = Arbitrary(customerGen(arbitrary[String]))