+++ scalding/src/main/scala/commbank/coppersmith/scalding/EavtText.scala
@@ -27,15 +27,16 @@ object EavtText {
   )
 
   implicit object EavtEnc extends FeatureValueEnc[Eavt] {
-    def encode(fvt: (FeatureValue[_], FeatureTime)): Eavt = fvt match {
+    def encode(fvt: (FeatureValue[Value], FeatureTime)): Eavt = fvt match {
       case (fv, time) =>
         val featureValue = (fv.value match {
           case Integral(v) => v.map(_.toString)
           case Decimal(v) => v.map(_.toString)
           case FloatingPoint(v) => v.map(_.toString)
           case Str(v) => v
-          case Date(v) => v.map(_.toString)
-          case Time(v) => v.map(_.toString)
+          case Bool(v) => v.map(_.toString)
+          case Date(v) => v.map(_.toIso8601ExtendedFormatString)
+          case Time(v) => v.map(_.toRfc3339String)
         }).getOrElse(HiveTextSink.NullValue)
 
         val featureTime = new DateTime(time).toString("yyyy-MM-dd")